# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Motionclash
  version: 1.0.0
servers:
- url: https://localhost:7000
paths:
  /api/user/{id}:
    delete:
      operationId: deleteUserById
      parameters:
      - description: "ID of the User to retrieve, update or delete."
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
      summary: Delete a User by ID.
      tags:
      - User
      x-accepts: application/json
      x-tags:
      - tag: User
    get:
      operationId: getUserById
      parameters:
      - description: "ID of the User to retrieve, update or delete."
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User object
        "404":
          description: User not found
      summary: Get a User by ID.
      tags:
      - User
      x-accepts: application/json
      x-tags:
      - tag: User
    put:
      operationId: upsertUserById
      parameters:
      - description: "ID of the User to retrieve, update or delete."
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: User object to upsert.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User object
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User created
        "404":
          description: User not found
      summary: Upsert a User by ID.
      tags:
      - User
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: User
  /api/user:
    get:
      operationId: getAllUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: List of User objects
      summary: Get all Users.
      tags:
      - User
      x-accepts: application/json
      x-tags:
      - tag: User
    post:
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: User object to create.
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User created
      summary: Create a new User.
      tags:
      - User
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: User
...
components:
  schemas:
    User:
      example:
        firstName: John
        lastName: James
        profilePicture:
          imagePath: imagePath
          id: 123e4567-e89b-12d3-a456-426614174000
          imageType: ProfilePicture
        password: "12345"
        phone: +36301234567
        id: 123e4567-e89b-12d3-a456-426614174000
        email: john@email.com
        username: theUser
      properties:
        id:
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        username:
          example: theUser
          type: string
        firstName:
          example: John
          type: string
        lastName:
          example: James
          type: string
        email:
          example: john@email.com
          type: string
        password:
          example: "12345"
          type: string
        phone:
          example: +36301234567
          type: string
        profilePicture:
          $ref: '#/components/schemas/Image'
      type: object
...
    Image:
      example:
        imagePath: imagePath
        id: 123e4567-e89b-12d3-a456-426614174000
        imageType: ProfilePicture
      properties:
        id:
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        imageType:
          enum:
          - ProfilePicture
          - EventPicture
          type: string
        imagePath:
          type: string
      type: object
...